Refactor unit.SetState into a ModelOperation
